improved userDetails loading to be independent of user-button component
