+++ apps/web/src/components/user-button/user-button.tsx
@@ -1,62 +1,21 @@
 'use client';
 
 import { Group, Avatar, Text, Flex } from '@mantine/core';
-import { useCallback, useEffect, useState, type ReactNode } from 'react';
+import { type ReactNode } from 'react';
 import { usePathname, useRouter } from 'next/navigation';
-import { useAtom } from 'jotai';
-import { logger } from '@repo/logger';
-import { notifications } from '@mantine/notifications';
-import { useTranslations } from 'next-intl';
-import { getUserDetails } from '@/app/(authenticated)/actions';
+import { useAtomValue } from 'jotai';
 import LoaderCentered from '@/components/misc/loader-centered';
 import { userDetailsAtom } from '@/app/atoms/user-atoms';
-import { IntegrationStatus, type Integration } from '@/graphql/generated/schema-server';
 import classes from './user-button.module.scss';
 
-export default function UserButton(): ReactNode {
-  const tGeneric = useTranslations('generic');
-  const tIntegrations = useTranslations('integrations');
+interface PropsType {
+  isDataLoaded: boolean;
+}
+
+export default function UserButton(props: PropsType): ReactNode {
   const router = useRouter();
   const pathname = usePathname();
-  const [userDetails, setUserDetails] = useAtom(userDetailsAtom);
-  const [isDataLoaded, setIsDataLoaded] = useState<boolean>(false);
-
-  const checkIntegrationTokensForExpiration = useCallback(
-    (integrationsData: Integration[] | null): void => {
-      if (integrationsData) {
-        for (const integration of integrationsData) {
-          if (integration.status === IntegrationStatus.Expiring) {
-            notifications.show({
-              title: tGeneric('warning'),
-              message: `(${integration.type}) ${tIntegrations('tokenWarning')}`,
-              color: 'orange',
-              autoClose: false,
-            });
-          } else if (integration.status === IntegrationStatus.Errored) {
-            notifications.show({
-              title: tGeneric('error'),
-              message: `(${integration.type}) ${tIntegrations('erroredIntegrationWarning')}`,
-              color: 'red',
-              autoClose: false,
-            });
-          }
-        }
-      }
-    },
-    [tGeneric, tIntegrations],
-  );
-
-  useEffect(() => {
-    void getUserDetails()
-      .then((res) => {
-        checkIntegrationTokensForExpiration(res.currentOrganization?.integrations as Integration[] | null);
-        setUserDetails(res);
-        setIsDataLoaded(true);
-      })
-      .catch((error: unknown) => {
-        logger.error(error);
-      });
-  }, [checkIntegrationTokensForExpiration, setUserDetails]);
+  const userDetails = useAtomValue(userDetailsAtom);
 
   const redirectToProfile = (): void => {
     router.push('profile');
@@ -67,7 +26,7 @@ export default function UserButton(): ReactNode {
       justify="flex-start"
       className={pathname === '/profile' ? `${classes.user} ${classes.buttonActive}` : classes.user}
     >
-      {isDataLoaded ? (
+      {props.isDataLoaded ? (
         <Group onClick={redirectToProfile}>
           <Avatar src={userDetails.photoUrl} radius="xl" />
