+++ apps/web/src/graphql/insights.graphql
@@ -13,6 +13,7 @@ query adAccounts {
 query insights(
   $adAccountIds: [String!]
   $adIds: [String!]
+  $integrations: [IntegrationType!]
   $dateFrom: Date
   $dateTo: Date
   $devices: [DeviceEnum!]
@@ -30,6 +31,7 @@ query insights(
     filter: {
       adAccountIds: $adAccountIds
       adIds: $adIds
+      integrations: $integrations
       dateFrom: $dateFrom
       dateTo: $dateTo
       devices: $devices
